Final pass
I think I've got it now. I'm now interpreting it as defaultValue instead
of initialValue, which helps a lot. I also added deprecation warnings
for <Broadcast> and <Subscriber> which should help people upgrade.

Only remaining question for me is: can I ship code that uses Symbol() to
the browser? Do I need a shim for that?